Model Preprocessing/requirements.txt
dlib==19.24.0
imutils==0.5.4
matplotlib==3.5.2
numpy==1.22.4
opencv_python==4.5.5.64
pandas==1.4.2
Pillow==9.1.1
tqdm==4.64.0
youtube_dl==2021.12.17
